fix(echo): audit burn-down — transcript corruption, security hardening, UX dead-ends
Text cleanup (silent transcript corruption on every dictation):
- sentence-start filler removal now requires a word boundary ("Solar
  panels" no longer loses its "So", "Welcome" its "Well")
- hallucination [..]/(..)/YouTube-phrase removal now matches whole
  utterances only - dictated parentheticals and phrases survive
- removed ambiguous words (may, march, sec, meta, apple) from the
  force-recase lists ("give me a sec" no longer becomes "a SEC")

Security:
- update downloads pinned to alanglobalintelligence.com (a compromised
  webview could previously supply any URL + matching hash = native RCE)
- transcribe/read_wav_base64 now enforce the same data-dir containment
  as discard_recording (webview could delete/read arbitrary files)

Reliability:
- trial: far-future clock ratchet (dead CMOS battery) no longer bricks
  the trial with a false "resets tomorrow"; small rollbacks stay locked
- self-update and EULA-decline quit now shut down the speech engine
  (process::exit skipped cleanup and orphaned it holding VRAM + locked
  files that could fail the NSIS install)
- GPU pack reinstall stops the engine before replacing the live pack dir
  (Windows refused to delete the running exe - install always failed)
- pack extraction preserves the unix executable bit (macOS pack was DOA)
- Vulkan pack no longer offered on macOS (Windows binary - guaranteed
  dead-end download on Intel Macs)
- pack install thread-spawn failure no longer wedges the progress
  channel for the session; db pagination offset no longer wraps u32

UI/UX:
- capability language only: GPU vendor names, CUDA/engine tech names,
  and raw model filenames no longer reach the UI
- language switch commits only after its model download succeeds
  (offline failure used to leave a language no installed model speaks)
- model-download "done" re-queries the backend instead of assuming the
  multilingual model landed; progress box now says which download it is
- update-error banner gains Retry + dismiss (was an undismissable dead-end)
- delete confirmation is an in-app two-step (window.confirm is a silent
  no-op on macOS - Delete did nothing there)
- StatusPanel Btn forwards `disabled` (onboarding's "Warming up" button
  was fully clickable); Enter can no longer double-submit license
  activation; gpu-pack progress listener no longer leaks on fast
  open/close; settings toggles roll back if the backend refuses the save

+++ src-tauri/src/db.rs
@@ -106,7 +106,9 @@ impl TranscriptDB {
 
     pub fn get_page(&self, page: u32, page_size: u32) -> Result<(Vec<Transcript>, i64)> {
         let total: i64 = self.conn.query_row("SELECT COUNT(*) FROM transcripts", [], |r| r.get(0))?;
-        let offset = page * page_size;
+        // u64 math: both values are frontend-controlled and u32 * u32 wraps
+        // in release builds, yielding a wrong OFFSET.
+        let offset = page as u64 * page_size as u64;
         let mut stmt = self.conn.prepare(
             "SELECT id, timestamp, text, raw_text, duration_seconds, character_count, word_count FROM transcripts ORDER BY timestamp DESC LIMIT ?1 OFFSET ?2"
         )?;
